pyproject.toml
[project]
name = "rlagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[groq]>=0.7.4",
    "pydantic-ai[logfire]>=0.7.4",
]
